refactor: split into packages under economy.
