--- src/main/kotlin/com/jtmnetwork/play/core/domain/constant/TransactionType.kt
@@ -1,5 +1,0 @@
-package com.jtmnetwork.play.core.domain.constant
-
-enum class TransactionType {
-    IN, OUT
-}
