+++ src/main/kotlin/com/jtmnetwork/play/core/domain/constant/economy/TransactionType.kt
@@ -1,0 +1,5 @@
+package com.jtmnetwork.play.core.domain.constant.economy
+
+enum class TransactionType {
+    IN, OUT
+}
