+++ src/main/kotlin/com/jtmnetwork/play/entrypoint/controller/economy/CurrencyController.kt
@@ -1,0 +1,36 @@
+package com.jtmnetwork.play.entrypoint.controller.economy
+
+import com.jtmnetwork.play.core.domain.entity.economy.Currency
+import com.jtmnetwork.play.data.service.economy.CurrencyService
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.web.bind.annotation.DeleteMapping
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.PutMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+import reactor.core.publisher.Flux
+import reactor.core.publisher.Mono
+import java.util.*
+
+@RestController
+@RequestMapping("/currency")
+class CurrencyController @Autowired constructor(private val currencyService: CurrencyService) {
+
+    @PostMapping
+    fun postCurrency(@RequestBody currency: Currency): Mono<Currency> = currencyService.insertCurrency(currency)
+
+    @PutMapping
+    fun putCurrency(@RequestBody currency: Currency): Mono<Currency> = currencyService.updateCurrency(currency)
+
+    @GetMapping("/{id}")
+    fun getCurrency(@PathVariable id: UUID): Mono<Currency> = currencyService.getCurrency(id)
+
+    @GetMapping("/all")
+    fun getCurrencies(): Flux<Currency> = currencyService.getCurrencies()
+
+    @DeleteMapping("/{id}")
+    fun deleteCurrency(@PathVariable id: UUID): Mono<Currency> = currencyService.removeCurrency(id)
+}
